unused import remove, reformat code

+++ src/main/java/companyOA/hackerrank/swiggy/BinaryManipulation.java
@@ -2,8 +2,6 @@ package companyOA.hackerrank.swiggy;
 
 import java.io.*;
 
-import static java.util.stream.Collectors.joining;
-
 
 class ResultBinaryManipulation {
 
@@ -15,19 +13,20 @@ class ResultBinaryManipulation {
      */
 
     public static long minOperations(long n) {
-    // Write your code here
+        // Write your code here
         int bit = n < 0 ? 1 : 0;
-        long res = (long)bit;
-        for(int i=0;i<63;i++) {
+        long res = bit;
+        for (int i = 0; i < 63; i++) {
             n = n << 1;
             int sign = n < 0 ? 1 : 0;
-            bit = (bit+sign)%2;
+            bit = (bit + sign) % 2;
             res = (res << 1) + bit;
         }
         return res;
     }
 
 }
+
 public class BinaryManipulation {
     public static void main(String[] args) throws IOException {
         BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
